Rename tmcore to core
