+++ src/maze/cell.rs
@@ -1,5 +1,5 @@
 use crate::maze::cell::CellWall::*;
-use crate::tmcore::*;
+use crate::core::*;
 
 #[derive(Clone)]
 pub struct Cell {
